wrap gcs sync client into async one, and replace the existing impl
