+++ python/ray/_private/gcs_utils.py
@@ -14,6 +14,8 @@ from ray.core.generated import (
     gcs_service_pb2_grpc,
 )
 
+import ray._private.gcs_aio_client
+
 from ray.core.generated.common_pb2 import ErrorType, JobConfig
 from ray.core.generated.gcs_pb2 import (
     ActorTableData,
@@ -180,198 +182,8 @@ class GcsCode(enum.IntEnum):
     GrpcUnavailable = 26
 
 
-class GcsAioClient:
-    def __init__(
-        self,
-        channel: Optional[GcsChannel] = None,
-        address: Optional[str] = None,
-        nums_reconnect_retry: int = 5,
-    ):
-        if channel is None:
-            assert isinstance(address, str)
-            channel = GcsChannel(gcs_address=address, aio=True)
-        assert isinstance(channel, GcsChannel)
-        assert channel._aio is True
-        self._channel = channel
-        self._connect()
-        self._nums_reconnect_retry = nums_reconnect_retry
-
-    @property
-    def channel(self):
-        return self._channel
-
-    def _connect(self):
-        self._channel.connect()
-        self._kv_stub = gcs_service_pb2_grpc.InternalKVGcsServiceStub(
-            self._channel.channel()
-        )
-        self._node_info_stub = gcs_service_pb2_grpc.NodeInfoGcsServiceStub(
-            self._channel.channel()
-        )
-        self._job_info_stub = gcs_service_pb2_grpc.JobInfoGcsServiceStub(
-            self._channel.channel()
-        )
-        self._actor_info_stub = gcs_service_pb2_grpc.ActorInfoGcsServiceStub(
-            self._channel.channel()
-        )
-
-    @_auto_reconnect
-    async def check_alive(
-        self, node_ips: List[bytes], timeout: Optional[float] = None
-    ) -> List[bool]:
-        req = gcs_service_pb2.CheckAliveRequest(raylet_address=node_ips)
-        reply = await self._node_info_stub.CheckAlive(req, timeout=timeout)
-
-        if reply.status.code != GcsCode.OK:
-            raise RuntimeError(
-                f"GCS running at {self._channel.address} is unhealthy: {reply.status}"
-            )
-        return list(reply.raylet_alive)
-
-    @_auto_reconnect
-    async def internal_kv_get(
-        self, key: bytes, namespace: Optional[bytes], timeout: Optional[float] = None
-    ) -> Optional[bytes]:
-        logger.debug(f"internal_kv_get {key!r} {namespace!r}")
-        req = gcs_service_pb2.InternalKVGetRequest(namespace=namespace, key=key)
-        reply = await self._kv_stub.InternalKVGet(req, timeout=timeout)
-        if reply.status.code == GcsCode.OK:
-            return reply.value
-        elif reply.status.code == GcsCode.NotFound:
-            return None
-        else:
-            raise RuntimeError(
-                f"Failed to get value for key {key!r} "
-                f"due to error {reply.status.message}"
-            )
-
-    @_auto_reconnect
-    async def internal_kv_multi_get(
-        self,
-        keys: List[bytes],
-        namespace: Optional[bytes],
-        timeout: Optional[float] = None,
-    ) -> Dict[bytes, bytes]:
-        logger.debug(f"internal_kv_multi_get {keys!r} {namespace!r}")
-        req = gcs_service_pb2.InternalKVMultiGetRequest(namespace=namespace, keys=keys)
-        reply = await self._kv_stub.InternalKVMultiGet(req, timeout=timeout)
-        if reply.status.code == GcsCode.OK:
-            return {entry.key: entry.value for entry in reply.results}
-        else:
-            raise RuntimeError(
-                f"Failed to get value for keys {keys!r} "
-                f"due to error {reply.status.message}"
-            )
-
-    @_auto_reconnect
-    async def internal_kv_put(
-        self,
-        key: bytes,
-        value: bytes,
-        overwrite: bool,
-        namespace: Optional[bytes],
-        timeout: Optional[float] = None,
-    ) -> int:
-        """Put a key-value pair into the GCS.
-
-        Args:
-            key: The key to put.
-            value: The value to put.
-            overwrite: Whether to overwrite the value if the key already exists.
-            namespace: The namespace to put the key-value pair into.
-            timeout: The timeout in seconds.
-
-        Returns:
-            The number of keys added. If overwrite is True, this will be 1 if the
-                key was added and 0 if the key was updated. If overwrite is False,
-                this will be 1 if the key was added and 0 if the key already exists.
-        """
-        logger.debug(f"internal_kv_put {key!r} {value!r} {overwrite} {namespace!r}")
-        req = gcs_service_pb2.InternalKVPutRequest(
-            namespace=namespace,
-            key=key,
-            value=value,
-            overwrite=overwrite,
-        )
-        reply = await self._kv_stub.InternalKVPut(req, timeout=timeout)
-        if reply.status.code == GcsCode.OK:
-            return reply.added_num
-        else:
-            raise RuntimeError(
-                f"Failed to put value {value!r} to key {key!r} "
-                f"due to error {reply.status.message}"
-            )
-
-    @_auto_reconnect
-    async def internal_kv_del(
-        self,
-        key: bytes,
-        del_by_prefix: bool,
-        namespace: Optional[bytes],
-        timeout: Optional[float] = None,
-    ) -> int:
-        logger.debug(f"internal_kv_del {key!r} {del_by_prefix} {namespace!r}")
-        req = gcs_service_pb2.InternalKVDelRequest(
-            namespace=namespace, key=key, del_by_prefix=del_by_prefix
-        )
-        reply = await self._kv_stub.InternalKVDel(req, timeout=timeout)
-        if reply.status.code == GcsCode.OK:
-            return reply.deleted_num
-        else:
-            raise RuntimeError(
-                f"Failed to delete key {key!r} " f"due to error {reply.status.message}"
-            )
-
-    @_auto_reconnect
-    async def internal_kv_exists(
-        self, key: bytes, namespace: Optional[bytes], timeout: Optional[float] = None
-    ) -> bool:
-        logger.debug(f"internal_kv_exists {key!r} {namespace!r}")
-        req = gcs_service_pb2.InternalKVExistsRequest(namespace=namespace, key=key)
-        reply = await self._kv_stub.InternalKVExists(req, timeout=timeout)
-        if reply.status.code == GcsCode.OK:
-            return reply.exists
-        else:
-            raise RuntimeError(
-                f"Failed to check existence of key {key!r} "
-                f"due to error {reply.status.message}"
-            )
-
-    @_auto_reconnect
-    async def internal_kv_keys(
-        self, prefix: bytes, namespace: Optional[bytes], timeout: Optional[float] = None
-    ) -> List[bytes]:
-        logger.debug(f"internal_kv_keys {prefix!r} {namespace!r}")
-        req = gcs_service_pb2.InternalKVKeysRequest(namespace=namespace, prefix=prefix)
-        reply = await self._kv_stub.InternalKVKeys(req, timeout=timeout)
-        if reply.status.code == GcsCode.OK:
-            return reply.results
-        else:
-            raise RuntimeError(
-                f"Failed to list prefix {prefix!r} "
-                f"due to error {reply.status.message}"
-            )
-
-    @_auto_reconnect
-    async def get_all_job_info(
-        self, timeout: Optional[float] = None
-    ) -> gcs_service_pb2.GetAllJobInfoReply:
-        req = gcs_service_pb2.GetAllJobInfoRequest()
-        reply = await self._job_info_stub.GetAllJobInfo(req, timeout=timeout)
-        return reply
-
-    @_auto_reconnect
-    async def get_named_actor_info(
-        self,
-        actor_name: str,
-        ray_namespace: str = "",
-        timeout: Optional[float] = None,
-    ) -> gcs_service_pb2.GetNamedActorInfoReply:
-        req = gcs_service_pb2.GetNamedActorInfoRequest(
-            name=actor_name, ray_namespace=ray_namespace
-        )
-        reply = await self._actor_info_stub.GetNamedActorInfo(req, timeout=timeout)
-        return reply
+# re-export
+GcsAioClient = ray._private.gcs_aio_client.GcsAioClient
 
 
 def cleanup_redis_storage(
